// packages/cardano-services/src/ChainHistory/DbSyncChainHistoryProvider/types.ts
import { Cardano } from '@cardano-sdk/core';

export type TransactionDataMap<T> = Map<Cardano.TransactionId, T>;
export type TxOutTokenMap = Map<string, Cardano.TokenMap>;
export type TxTokenMap = TransactionDataMap<Cardano.TokenMap>;

export interface BlockModel {
  block_no: number;
  epoch_no: number;
  epoch_slot_no: number;
  hash: Buffer;
  next_block: Buffer | null;
  previous_block: Buffer | null;
  size: number;
  slot_leader_hash: Buffer;
  slot_leader_pool: string | null;
  slot_no: number;
  time: string;
  tx_count: string;
  vrf: string;
}

export interface BlockOutputModel {
  fees: string;
  output: string;
  hash: Buffer;
}

export interface TipModel {
  block_no: number;
  hash: Buffer;
  slot_no: number;
}

export interface TxModel {
  id: Buffer;
  index: number;
  size: number;
  fee: string;
  invalid_before: string | null;
  invalid_hereafter: string | null;
  block_no: number;
  block_hash: Buffer;
  block_slot_no: number;
}

export interface TxInOutModel {
  address: string;
  coin_value: string;
  datum?: Buffer | null;
  id: string;
  index: number;
  tx_id: Buffer;
}

export interface TxOutput extends Cardano.TxOut {
  txId: Cardano.TransactionId;
  index: number;
}

export interface MultiAssetModel {
  asset_name: Buffer;
  fingerprint: string;
  policy_id: Buffer;
  quantity: string;
  tx_id: Buffer;
}

export interface TxOutMultiAssetModel extends MultiAssetModel {
  tx_out_id: string;
}

export interface WithdrawalModel {
  quantity: string;
  tx_id: Buffer;
  stake_address: string;
}

export interface RedeemerModel {
  index: number;
  purpose: string;
  script_hash: Buffer;
  unit_mem: string;
  unit_steps: string;
  tx_id: Buffer;
}

export interface TxMetadataModel {
  key: string;
  json_value: { [k: string]: unknown };
  tx_id: Buffer;
}

export interface CertificateModel {
  cert_index: number;
  tx_id: Buffer;
}
export type WithCertIndex<T extends Cardano.Certificate> = T & { cert_index: number };
export type WithCertType<T extends CertificateModel> = T & {
  type: 'retire' | 'update' | 'mir' | 'stake' | 'delegation';
};

export interface PoolRetireCertModel extends CertificateModel {
  retiring_epoch: number;
  pool_id: string;
}
export const isPoolRetireCertModel = (
  value: WithCertType<CertificateModel>
): value is WithCertType<PoolRetireCertModel> => value.type === 'retire';

export interface PoolUpdateCertModel extends CertificateModel {
  pool_id: string;
}
export const isPoolUpdateCertModel = (
  value: WithCertType<CertificateModel>
): value is WithCertType<PoolUpdateCertModel> => value.type === 'update';

export interface MirCertModel extends CertificateModel {
  amount: string;
  pot: 'reserve' | 'treasury';
  address: string;
}
export const isMirCertModel = (value: WithCertType<CertificateModel>): value is WithCertType<MirCertModel> =>
  value.type === 'mir';

export interface StakeCertModel extends CertificateModel {
  address: string;
  registration: boolean;
}
export const isStakeCertModel = (value: WithCertType<CertificateModel>): value is WithCertType<StakeCertModel> =>
  value.type === 'stake';

export interface DelegationCertModel extends CertificateModel {
  address: string;
  pool_id: string;
}
export const isDelegationCertModel = (
  value: WithCertType<CertificateModel>
): value is WithCertType<DelegationCertModel> => value.type === 'delegation';

export interface ProtocolParamsModel {
  coin_per_utxo_word: string;
  max_collateral_inputs: number;
  max_tx_size: number;
  max_val_size: string;
  min_fee_coefficient: number;
  min_fee_constant: number;
  min_pool_cost: string;
  pool_deposit: string;
  protocol_major: number;
  protocol_minor: number;
  key_deposit: string;
}